Extracted the JPanel into separate a file
It was an inline anonymous class, needed more clarity.
